# src/infrastructure/openapi/definitions/openapi-v1.yml
openapi: 3.0.3
info:
  title: I Was Here Service API
  version: 1.0.0
  description: I Was Here Service API built with NestJS
servers:
  - url: http://localhost:3000
    description: Development server
paths:
...
  /posts:
    post:
      operationId: savePost
      summary: Create a new post
      description: Create a new post with location data, text content, and optional media attachments
      security:
        - bearerAuth: []
      tags:
        - Posts
      parameters:
        - name: latitude
          in: header
          required: true
          description: User's current latitude for post location
          schema:
            type: number
            format: double
            minimum: -90
            maximum: 90
          example: 40.7589
        - name: longitude
          in: header
          required: true
          description: User's current longitude for post location
          schema:
            type: number
            format: double
            minimum: -180
            maximum: 180
          example: -73.9851
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#components/schemas/SavePostRequest'
      responses:
        '201':
          description: Post successfully created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SavePostResponse'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/JwtUnauthorizedError'
        '422':
          $ref: '#/components/responses/UnprocessableEntityError'
  /posts/bounding-box:
    get:
      operationId: findPostsInBoundingBox
      summary: Find posts within a bounding box
      description: Retrieve posts within a specified rectangular area defined by center coordinates and dimensions
      security:
        - bearerAuth: []
      tags:
        - Posts
      parameters:
        - name: latitude
          in: header
          required: true
          description: Center latitude of the bounding box
          schema:
            type: number
            format: double
            minimum: -90
            maximum: 90
          example: 40.7128
        - name: longitude
          in: header
          required: true
          description: Center longitude of the bounding box
          schema:
            type: number
            format: double
            minimum: -180
            maximum: 180
          example: -74.0060
        - name: box-length
          in: query
          required: true
          description: Length of the bounding box in longitude degrees
          schema:
            type: number
            format: double
            minimum: 0.001
            maximum: 0.1
          example: 0.01
        - name: box-width
          in: query
          required: true
          description: Width of the bounding box in latitude degrees
          schema:
            type: number
            format: double
            minimum: 0.001
            maximum: 0.1
          example: 0.01
      responses:
        '200':
          description: Posts within bounding box successfully retrieved
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/FindPostsInBoundingBoxResponse'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/JwtUnauthorizedError'
        '422':
          $ref: '#/components/responses/UnprocessableEntityError'
  /posts/vicinity:
    get:
      operationId: findPostsInVicinity
      summary: Find posts in user's vicinity
      description: Retrieve posts within 15 meters of the user's current location, sorted by approval count with engagement metrics
      security:
        - bearerAuth: []
      tags:
        - Posts
      parameters:
        - name: latitude
          in: header
          required: true
          description: User's current latitude
          schema:
            type: number
            format: double
            minimum: -90
            maximum: 90
          example: 40.7128
        - name: longitude
          in: header
          required: true
          description: User's current longitude
          schema:
            type: number
            format: double
            minimum: -180
            maximum: 180
          example: -74.0060
        - name: page
          in: query
          required: false
          description: Page number for pagination (default is 1)
          schema:
            type: integer
            minimum: 1
            default: 1
          example: 1
      responses:
        '200':
          description: Posts in vicinity successfully retrieved
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/FindPostsInVicinityResponse'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/JwtUnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '422':
          $ref: '#/components/responses/UnprocessableEntityError'
  /posts/{uuid}:
    get:
      operationId: findPostByUuid
      summary: Get post details by UUID
      description: Retrieve a post with its creator information, comments, and approval count by providing the user's current location to verify they are within proximity
      security:
        - bearerAuth: []
      tags:
        - Posts
      parameters:
        - name: uuid
          in: path
          required: true
          description: UUID of the post to retrieve
          schema:
            type: string
            format: uuid
          example: 123e4567-e89b-12d3-a456-426614174002
        - name: latitude
          in: header
          required: true
          description: User's current latitude for location verification
          schema:
            type: number
            format: double
            minimum: -90
            maximum: 90
          example: 40.7589
        - name: longitude
          in: header
          required: true
          description: User's current longitude for location verification
          schema:
            type: number
            format: double
            minimum: -180
            maximum: 180
          example: -73.9851
      responses:
        '200':
          description: Post details successfully retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FindPostByUuidResponse'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/JwtUnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '422':
          $ref: '#/components/responses/UnprocessableEntityError'
...
components:
  securitySchemes: 
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    refreshToken:
      type: apiKey
      in: header
      name: refresh-token
      description: Refresh token for renewing JWT
  responses:
    BadRequestError:
      description: Missing required parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            message: Required parameters are missing
            error: Bad Request
            statusCode: 400
    JwtUnauthorizedError:
      description: Failure to provide a valid JWT token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            message: 'Authentication Failed'
            error: 'Unauthorized'
            statusCode: 401
...
    UnprocessableEntityError:
      description: Provided parameters are invalid
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            message: Failed to validate parameters
            error: Unprocessable Entity
            statusCode: 422
...
    ForbiddenError:
      description: User is not authorized to perform this action or does not meet location requirements
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            message: 'User is not near post'
            error: 'Forbidden'
            statusCode: 403
    NotFoundError:
      description: The requested resource was not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            message: 'Not Found'
            error: 'Not Found'
            statusCode: 404 
  schemas:
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
          description: Message describing the error
        error:
          type: string
          description: Error type
        statusCode:
          type: integer
          description: HTTP status status code
...
    SavePostResponse:
      type: object
      properties:
        post:
          type: object
          properties:
            uuid:
              type: string
              format: uuid
              description: Post UUID
              example: 123e4567-e89b-12d3-a456-426614174002
            title:
              type: string
              description: Post title
              example: Beautiful sunset
            location:
              type: string
              description: Human-readable location description
              example: Central Park, New York
            latitude:
              type: number
              format: float
              description: Post location latitude
              example: 40.7589
            longitude:
              type: number
              format: float
              description: Post location longitude
              example: -73.9851
            body:
              type: string
              description: Post content
              example: Amazing view from the top of the mountain
            imageName:
              type: string
              nullable: true
              description: Image filename
              example: sunset_photo_001.jpg
            videoName:
              type: string
              nullable: true
              description: Video filename
              example: sunset_video_001.mp4
            createdAt:
              type: string
              format: date-time
              description: Post creation timestamp
              example: 2025-11-24T10:30:00Z
            updatedAt:
              type: string
              format: date-time
              description: Post last update timestamp
              example: 2025-11-24T10:30:00Z
            deletedAt:
              type: string
              nullable: true
              format: date-time
              description: Post soft deletion timestamp
              example: null
      required:
        - post
...
    FindPostByUuidResponse:
      type: object
      properties:
        post:
          type: object
          description: Found post
          properties:
            uuid:
              type: string
              format: uuid
              description: Post UUID
              example: 123e4567-e89b-12d3-a456-426614174002
            location:
              type: string
              description: Human-readable location description
              example: Central Park, New York
            latitude:
              type: number
              format: float
              description: Post location latitude
              example: 40.7589
            longitude:
              type: number
              format: float
              description: Post location longitude
              example: -73.9851
            title:
              type: string
              description: Post title
              example: Beautiful sunset
            body:
              type: string
              description: Post content
              example: Amazing view from the top of the mountain
            imageName:
              type: string
              nullable: true
              description: Image filename
              example: sunset_photo_001.jpg
            videoName:
              type: string
              nullable: true
              description: Video filename
              example: sunset_video_001.mp4
            createdAt:
              type: string
              format: date-time
              description: Post creation timestamp
              example: 2025-11-24T10:30:00Z
            updatedAt:
              type: string
              format: date-time
              description: Post last update timestamp
              example: 2025-11-24T10:30:00Z
            deletedAt:
              type: string
              nullable: true
              format: date-time
              description: Post soft deletion timestamp
              example: null
            isApproved:
              type: boolean
              description: Whether the requesting user has approved this post
              example: true
        creator:
          type: object
          description: User profile of the post creator
          properties:
            uuid:
              type: string
              format: uuid
              description: User profile UUID
              example: 987e4567-e89b-12d3-a456-426614174999
            nickname:
              type: string
              description: User's display nickname
              example: JohnDoe2024
            profilePictureName:
              type: string
              nullable: true
              description: Profile picture filename
              example: profile_pic_001.jpg
        comments:
          type: array
          description: List of comments on the post
          items:
            type: object
            properties:
              uuid:
                type: string
                format: uuid
                description: Comment UUID
                example: 456e4567-e89b-12d3-a456-426614174777
              body:
                type: string
                description: Comment content
                example: Great photo!
              createdAt:
                type: string
                format: date-time
                description: Comment creation timestamp
                example: 2025-11-24T12:00:00Z
              updatedAt:
                type: string
                format: date-time
                description: Comment last update timestamp
                example: 2025-11-24T12:00:00Z
              user:
                type: object
                description: User profile of the comment author
                properties:
                  uuid:
                    type: string
                    format: uuid
                    description: User profile UUID
                    example: 111e4567-e89b-12d3-a456-426614174888
                  nickname:
                    type: string
                    description: User's display nickname
                    example: Jane Smith
                  profilePictureName:
                    type: string
                    nullable: true
                    description: Profile picture filename
                    example: profile_pic_002.jpg
        approvalCount:
          type: integer
          description: Number of approvals for this post
          example: 15
...
    FindPostsInVicinityResponse:
      type: object
      properties:
        post:
          type: object
          properties:
            uuid:
              type: string
              format: uuid
              description: Post UUID
              example: 123e4567-e89b-12d3-a456-426614174000
            title:
              type: string
              description: Post title
              example: Beautiful sunset
            body:
              type: string
              description: Post content
              example: Amazing view from the park
            location:
              type: string
              description: Human-readable location description
              example: Washington Square Park, NYC
            latitude:
              type: number
              format: float
              description: Post location latitude
              example: 40.7128
            longitude:
              type: number
              format: float
              description: Post location longitude
              example: -74.0060
            imageName:
              type: string
              nullable: true
              description: Image filename
              example: sunset.jpg
            videoName:
              type: string
              nullable: true
              description: Video filename
              example: sunset_video.mp4
            createdAt:
              type: string
              format: date-time
              description: Post creation timestamp
              example: 2025-11-25T10:30:00Z
            updatedAt:
              type: string
              format: date-time
              description: Post last update timestamp
              example: 2025-11-25T10:30:00Z
            deletedAt:
              type: string
              nullable: true
              format: date-time
              description: Post soft deletion timestamp
              example: null
            isApproved:
              type: boolean
              description: Whether the requesting user has approved this post
              example: false
        creator:
          type: object
          description: Post creator information
          properties:
            uuid:
              type: string
              format: uuid
              description: User UUID
              example: user-uuid-123
            username:
              type: string
              description: Username
              example: johndoe
            profile:
              type: object
              properties:
                uuid:
                  type: string
                  format: uuid
                  description: User profile UUID
                  example: profile-uuid-123
                nickname:
                  type: string
                  description: User's display nickname
                  example: John Doe
                profilePictureName:
                  type: string
                  nullable: true
                  description: Profile picture filename
                  example: john.jpg
                points:
                  type: integer
                  description: User's points
                  example: 150
        approvalCount:
          type: integer
          description: Number of approvals for this post
          example: 25
        commentCount:
          type: integer
          description: Number of comments on this post
          example: 8
...
    FindPostsInBoundingBoxResponse:
      type: object
      properties:
        post:
          type: object
          properties:
            uuid:
              type: string
              format: uuid
              description: Post UUID
              example: 123e4567-e89b-12d3-a456-426614174000
            title:
              type: string
              description: Post title
              example: Beautiful sunset
            body:
              type: string
              description: Post content
              example: Amazing view from the park
            location:
              type: string
              description: Human-readable location description
              example: Brooklyn Bridge Park, NYC
            latitude:
              type: number
              format: float
              description: Post location latitude
              example: 40.7128
            longitude:
              type: number
              format: float
              description: Post location longitude
              example: -74.0060
            imageName:
              type: string
              nullable: true
              description: Image filename
              example: sunset.jpg
            videoName:
              type: string
              nullable: true
              description: Video filename
              example: sunset_video.mp4
            createdAt:
              type: string
              format: date-time
              description: Post creation timestamp
              example: 2025-11-25T10:30:00Z
            updatedAt:
              type: string
              format: date-time
              description: Post last update timestamp
              example: 2025-11-25T10:30:00Z
            deletedAt:
              type: string
              nullable: true
              format: date-time
              description: Post soft deletion timestamp
              example: null
        creator:
          type: object
          description: Post creator information
          properties:
            uuid:
              type: string
              format: uuid
              description: User UUID
              example: user-uuid-123
            profile:
              type: object
              properties:
                uuid:
                  type: string
                  format: uuid
                  description: User profile UUID
                  example: profile-uuid-123
                nickname:
                  type: string
                  description: User's display nickname
                  example: John Doe
                profilePictureName:
                  type: string
                  nullable: true
                  description: Profile picture filename
                  example: john.jpg
                points:
                  type: integer
                  description: User's points
                  example: 150
...