# src/posts/dto/find-post-by-uuid-response.dto.ts
export class FindPostByUuidResponseDto {
  post: {
    uuid: string;
    title: string;
    body: string;
    location: string;
    latitude: number;
    longitude: number;
    imageName: string;
    videoName: string;
    createdAt: Date;
    updatedAt: Date;
    deletedAt: Date;
    isApproved: boolean;
  };

  creator: {
    uuid: string;
    profile: {
      uuid: string;
      nickname: string;
      profilePictureName: string;
      points: number;
    };
  };

  comments: {
    uuid: string;
    body: string;
    createdAt: Date;
    updatedAt: Date;
    deletedAt: Date;
    user: {
      uuid: string;
      profile: {
        uuid: string;
        nickname: string;
        profilePictureName: string;
        points: number;
      };
    };
  }[];

  approvalCount: number;
}